+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ### 0.6.0  Dark Theme and History Display (WIP)
 - Changes UI to use new dark theme
 - #20 Adds display of navigation history for Back/Forward (Thanks @jilleJr)
+- #17 Adds display of undo/redo history
 - #63 Adds per-folder storage of sort settings (Thanks @jilleJr)
 - #94 Adds support for environment variable substitution in path box (Thanks @jilleJr)
 - #99 Adds repeat count for cursor movement, Parent Folder, Back, Forward, Undo, Redo, Find Next (Thanks @jilleJr)